PIM-7251: add behat tests
