+++ features/category/display_category_history.feature
@@ -17,9 +17,9 @@ Feature: Display the category history
     When I visit the "History" tab
     Then there should be 1 update
     And I should see history:
-      | version | property    | value         |
-      | 1       | code        | book          |
-      | 1       | label-en_US | Book category |
+      | version | property    | value         | date |
+      | 1       | code        | book          | now  |
+      | 1       | label-en_US | Book category | now  |
     When I visit the "Properties" tab
     And I fill in the following information:
       | English (United States) | My book category |
@@ -28,7 +28,7 @@ Feature: Display the category history
     When I visit the "History" tab
     Then there should be 2 updates
     And I should see history:
-      | version | property    | value            |
-      | 1       | code        | book             |
-      | 1       | label-en_US | Book category    |
-      | 2       | label-en_US | My book category |
+      | version | property    | value            | date |
+      | 1       | code        | book             | now  |
+      | 1       | label-en_US | Book category    | now  |
+      | 2       | label-en_US | My book category | now  |
