+++ features/channel/display_channel_history.feature
@@ -19,9 +19,9 @@ Feature: Display the channel history
     And I visit the "History" tab
     Then there should be 1 update
     And I should see history:
-      | version | property    | value           |
-      | 1       | code        | foo             |
-      | 1       | label-en_US | bar             |
-      | 1       | category    | 2014_collection |
-      | 1       | currencies  | EUR             |
-      | 1       | locales     | fr_FR           |
+      | version | property    | value           | date |
+      | 1       | code        | foo             | now  |
+      | 1       | label-en_US | bar             | now  |
+      | 1       | category    | 2014_collection | now  |
+      | 1       | currencies  | EUR             | now  |
+      | 1       | locales     | fr_FR           | now  |
