+++ features/localization/product/display_localized_history.feature
@@ -33,13 +33,13 @@ Feature: Display the localized product history
     When I visit the "Historique" column tab
     Then there should be 1 update
     And I should see history:
-      | version | property    | value     |
-      | 1       | SKU         | boots     |
-      | 1       | Metric      | 12,1234   |
-      | 1       | Metric unit | Gramme    |
-      | 1       | Number      | 98,7654   |
-      | 1       | Price EUR   | 20,80 €   |
-      | 1       | Price USD   | 25,35 $US |
+      | version | property    | value     | date |
+      | 1       | SKU         | boots     | now  |
+      | 1       | Metric      | 12,1234   | now  |
+      | 1       | Metric unit | Gramme    | now  |
+      | 1       | Number      | 98,7654   | now  |
+      | 1       | Price EUR   | 20,80 €   | now  |
+      | 1       | Price USD   | 25,35 $US | now  |
 
   Scenario: Display english-format product history numbers
     Given I am logged in as "Julia"
@@ -49,10 +49,10 @@ Feature: Display the localized product history
     When I visit the "History" column tab
     Then there should be 1 update
     And I should see history:
-      | version | property    | value   |
-      | 1       | SKU         | boots   |
-      | 1       | Metric      | 12.1234 |
-      | 1       | Metric unit | Gram    |
-      | 1       | Number      | 98.7654 |
-      | 1       | Price EUR   | €20.80  |
-      | 1       | Price USD   | $25.35  |
+      | version | property    | value   | date |
+      | 1       | SKU         | boots   | now  |
+      | 1       | Metric      | 12.1234 | now  |
+      | 1       | Metric unit | Gram    | now  |
+      | 1       | Number      | 98.7654 | now  |
+      | 1       | Price EUR   | €20.80  | now  |
+      | 1       | Price USD   | $25.35  | now  |
