+++ features/mass-action/mass_edit_and_update_history_family.feature
@@ -25,14 +25,14 @@ Feature: Update product history when mass editing products
     And I visit the "History" column tab
     Then there should be 2 updates
     And I should see history:
-      | version | property | value   |
-      | 2       | family   | sandals |
+      | version | property | value   | date |
+      | 2       | family   | sandals | now  |
     When I edit the "sneakers" product
     And I visit the "History" column tab
     Then there should be 2 updates
     And I should see history:
-      | version | property | value   |
-      | 2       | family   | sandals |
+      | version | property | value   | date |
+      | 2       | family   | sandals | now  |
     When I edit the "sandals" product
     And I visit the "History" column tab
     Then there should be 1 update
