+++ features/mass-action/mass_edit_and_update_history_status.feature
@@ -43,17 +43,17 @@ Feature: Update product history when mass editing products
     And I visit the "History" column tab
     Then there should be 2 updates
     And I should see history:
-      | version | property | value |
-      | 2       | enabled  | 0     |
+      | version | property | value | date |
+      | 2       | enabled  | 0     | now  |
     When I edit the "sandals" product
     And I visit the "History" column tab
     Then there should be 2 updates
     And I should see history:
-      | version | property | value |
-      | 2       | enabled  | 0     |
+      | version | property | value | date |
+      | 2       | enabled  | 0     | now  |
     When I edit the "sneakers" product
     And I visit the "History" column tab
     Then there should be 2 updates
     And I should see history:
-      | version | property | value |
-      | 2       | enabled  | 0     |
+      | version | property | value | date |
+      | 2       | enabled  | 0     | now  |
