+++ features/product/history_update.feature
@@ -24,17 +24,17 @@ Feature: Update the product history
     When I visit the "History" column tab
     Then there should be 2 update
     And I should see history:
-      | version | property  | value  |
-      | 2       | Price EUR | €10.00 |
-      | 2       | Price USD | $20.00 |
+      | version | property  | value  | date |
+      | 2       | Price EUR | €10.00 | now  |
+      | 2       | Price USD | $20.00 | now  |
     When I visit the "Attributes" column tab
     And I change the "Price" to "19 USD"
     And I save the product
     And I visit the "History" column tab
     Then there should be 3 updates
     And I should see history:
-      | version | property  | value  |
-      | 3       | Price USD | $19.00 |
+      | version | property  | value  | date |
+      | 3       | Price USD | $19.00 | now  |
 
   @jira https://akeneo.atlassian.net/browse/PIM-3628
   Scenario: Update product history when updating product metric
@@ -44,17 +44,17 @@ Feature: Update the product history
     When I visit the "History" column tab
     Then there should be 2 update
     And I should see history:
-      | version | property                 | value   |
-      | 2       | Washing temperature      | 40      |
-      | 2       | Washing temperature unit | Celsius |
+      | version | property                 | value   | date |
+      | 2       | Washing temperature      | 40      | now  |
+      | 2       | Washing temperature unit | Celsius | now  |
     When I visit the "Attributes" column tab
     And I change the "Washing temperature" to "35 Celsius"
     And I save the product
     When I visit the "History" column tab
     Then there should be 3 updates
     And I should see history:
-      | version | property            | value |
-      | 3       | Washing temperature | 35    |
+      | version | property            | value | date |
+      | 3       | Washing temperature | 35    | now  |
 
   @jira https://akeneo.atlassian.net/browse/PIM-3628
   Scenario: Update product history when updating product media
@@ -64,5 +64,5 @@ Feature: Update the product history
     When I visit the "History" column tab
     Then there should be 2 updates
     And I should see history:
-      | version | property  | value           |
-      | 2       | Thumbnail | .*SNKRS_1R\.png |
+      | version | property  | value           | date |
+      | 2       | Thumbnail | .*SNKRS_1R\.png | now  |
