+++ src/Pim/Bundle/DataGridBundle/Extension/Formatter/Property/ProductValue/DateProperty.php
@@ -38,19 +38,4 @@ class DateProperty extends FieldProperty
 
         return $this->presenter->present($result, ['locale' => $this->translator->getLocale()]);
     }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getRawValue(ResultRecordInterface $record)
-    {
-        try {
-            $value = $record->getValue($this->getOr(self::DATA_NAME_KEY, $this->get(self::NAME_KEY)));
-        } catch (\LogicException $e) {
-            // default value
-            $value = null;
-        }
-
-        return $value;
-    }
 }
