+++ src/Pim/Bundle/DataGridBundle/Resources/config/formatters.yml
@@ -31,13 +31,13 @@ services:
         tags:
             - { name: oro_datagrid.extension.formatter.property, type: product_value_date }
 
-    pim_datagrid.extension.formatter.property.product_value.datetime_property:
-        class: '%pim_datagrid.extension.formatter.property.product_value.date_property.class%'
+    pim_datagrid.extension.formatter.property.datetime_property:
+        class: '%pim_datagrid.extension.formatter.property.datetime_property.class%'
         arguments:
             - '@translator'
             - '@pim_catalog.localization.presenter.datetime'
         tags:
-            - { name: oro_datagrid.extension.formatter.property, type: product_value_datetime }
+            - { name: oro_datagrid.extension.formatter.property, type: property_date }
 
     pim_datagrid.extension.formatter.property.changeset:
         class: '%pim_datagrid.extension.formatter.property.changeset.class%'
