+++ src/Pim/Bundle/EnrichBundle/Resources/public/templates/product/history.html
@@ -5,7 +5,7 @@
                 <th class="AknGrid-headerCell"></th>
                 <th class="AknGrid-headerCell version"><%- _.__('pim_enrich.entity.product.history.version') %></th>
                 <th class="AknGrid-headerCell author"><%- _.__('pim_enrich.entity.product.history.author') %></th>
-                <th class="AknGrid-headerCell logged-at"><%- _.__('pim_enrich.entity.product.history.logged_at') %></th>
+                <th class="AknGrid-headerCell loggedAt"><%- _.__('pim_enrich.entity.product.history.logged_at') %></th>
                 <th class="AknGrid-headerCell changes"><%- _.__('pim_enrich.entity.product.history.modified') %></th>
                 <% if (hasAction) { %><th class="AknGrid-headerCell actions"><%- _.__('pim_enrich.entity.product.history.actions') %></th><% } %>
             </tr>
@@ -18,7 +18,7 @@
                     </td>
                     <td class="AknGrid-bodyCell version" data-column="version"><span class="AknButton AknButton--grey"><%- version.version %></span></td>
                     <td class="AknGrid-bodyCell AknGrid-bodyCell--noWrap AknGrid-bodyCell--highlight author" data-column="author"><%- version.author %><%- version.context ? ' (' + version.context + ')' : '' %></td>
-                    <td class="AknGrid-bodyCell AknGrid-bodyCell--noWrap logged-at" data-column="logged-at"><%- version.logged_at %></td>
+                    <td class="AknGrid-bodyCell AknGrid-bodyCell--noWrap loggedAt" data-column="loggedAt"><%- version.logged_at %></td>
                     <td class="AknGrid-bodyCell AknGrid-bodyCell--truncatable changes" data-column="changes"><div class="AknGrid-truncatable" title="<%- _.keys(version.changeset).join(', ') %>"><%- _.keys(version.changeset).join(', ') %></div></td>
                     <% if (hasAction) { %><td class="AknGrid-bodyCell actions"></td><% } %>
                 </tr>
